pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dagscheduler-env"
version = "1.0.0"
description = "A DAG-based job scheduling OpenEnv environment"
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "openai",
    "pyyaml",
    "python-dotenv",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["environment*", "tasks*"]